move Eye class to body.py file
